Remove integration tests
